few fixes

+++ entrypoint.sh
@@ -28,20 +28,23 @@ fi
 ############## Download and extract the latest OpenGrep rules from GitHub: END ############## 
 
 # Run opengrep with CMD arguments passed from Docker
+echo opengrep scan . -f "$RULES_DIR" --json --output results.json "$@"
 opengrep scan . -f "$RULES_DIR" --json --output results.json "$@"
 
-# Append metadata to the results JSON file using jq
-jq --arg repo "$REPO_NAME" \
-   --arg sha "$COMMIT_SHA" \
-   --arg ref "$COMMIT_REF" \
-   --arg run_id "$PIPELINE_ID" \
-   --arg repo_url "$REPOSITORY_URL" \
-   --arg repo_run_url "$JOB_URL" \
-   '. + {
-     repo: $repo,
-     sha: $sha,
-     ref: $ref,
-     run_id: $run_id,
-     repo_url: $repo_url,
-     repo_run_url: $repo_run_url
-   }' results.json > tmp.json && mv tmp.json results.json
+if [[ -f results.json ]]; then
+  # Append metadata to the results JSON file using jq
+  jq --arg repo "$REPO_NAME" \
+    --arg sha "$COMMIT_SHA" \
+    --arg ref "$COMMIT_REF" \
+    --arg run_id "$PIPELINE_ID" \
+    --arg repo_url "$REPOSITORY_URL" \
+    --arg repo_run_url "$JOB_URL" \
+    '. + {
+      repo: $repo,
+      sha: $sha,
+      ref: $ref,
+      run_id: $run_id,
+      repo_url: $repo_url,
+      repo_run_url: $repo_run_url
+    }' results.json > tmp.json && mv tmp.json results.json
+fi
